src/emr_jobs/pyproject.toml
[tool.poetry]
name = "emr-predictive-model"
version = "0.1.0"
description = "PySpark job to run predictive models on stock tickers"
authors = ["Josh Phillips <you@example.com>"]
packages = [{ include = "emr_jobs" }]

[tool.poetry.dependencies]
python = "^3.12"
pyspark = "^3.2.1"
boto3 = "^1.26.0"
psycopg2-binary = "^2.9.3"
scikit-learn = "^1.0.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"